separate confidence interval class to separate module
